Overhauled IRemoteFileStoreConnection interface
This interface now supports all of the currently important operations
with both file paths (which shouldn't be used often) and from item ids.
It should be noted that this is due to OneDrive's specific way of saving
files.  However, it is likely that onedrive and other cloud services use
a similar feature.  In any other case, the "id" could be the same as the
"remote path" in implementing the interface.