Restructured local file event handling
now local file event handling is a queue and attempts to deal with
network connections failing by refusing to process the event queue if
the next item fails (see #15).  In the real world, this is bad design
and these items should be set aside for the user to deal with.  I don't
even really know what else could cause this.