Significantly changed sync structure
the ActiveSyncFileStore was becoming monolithic, so there are a few new
layers to split up its responsibility.  First there is the
BufferedRemoteFileStoreInterface, which buffers requests to the remote
before sending them, allowing for synchronous request submission and
status/progress callback.  The LocalFileStoreInterface takes the events
from the ILocalFileStore and interprets them into a more usable form
while filtering unnecessary ones.  The FileStoreBridge is the layer
between the two former with rather simple logic.