Improved moved-item handling and downloading logic
Metadata doesn't get modified until after a move request is detected, so
delete/create events don't prematurely remove important metadata.  This
gets modified in the `GetDeltas` method of `LocalFileStoreInterface`.
Next, the logic used to request downloads from the `FileStoreBridge`
listens to the `BufferedRemoteFileStoreInterface` to wait for the
download to be complete before setting the `LastModified` of the local
file to the remote.