Lots of work on communication and local queue
the local queue processing is much more fleshed-out, but some of the
nitty-gritty is not quite done yet.  There is still a major issue with
communicating the last modified from the remote to the local when
uploading a file.  This will result in the just-uploaded file needing to
have its sha1 sum calculated unncessarily.