Delete request checks against remote last modified
this address issue #30.  By using the most recient known last modified
of a remote file, the remote delete request can be sent to the local
with a last modified timestamp.  This means if the local file has been
modified since the delete took place, then we can rebound it to the
remote.